Rename functions to follow the dev branch's setup
