+++ src/functions/update.rs
@@ -1,67 +1,46 @@
 use crate::errors;
-use crate::models::Config;
 use clap::ArgMatches;
 use colored::*;
-use git2::{Oid, Repository};
-use log::{error, trace};
-pub fn update(_args: &ArgMatches) -> Result<(), errors::LeftError> {
-    let mut config = Config::load().unwrap_or_default();
-    //attempt to fetch new themes
-    println!("    Retrieving themes from {:?}", &config.source());
-    let resp = reqwest::blocking::get(&config.source())?.text_with_charset("utf-8")?;
-    trace!("{:?}", &resp);
+use log::trace;
 
-    //compare to old themes
-    config.compare(toml::from_str(&resp)?)?;
-    trace!("{:?}", &config);
-    Config::save(&config)?;
-    println!("{}", "\nUpdating themes:".blue().bold());
-    let mut installed = 0;
+pub fn update(args: &ArgMatches) -> Result<(), errors::LeftError> {
+    println!("{}", "Fetchning themes . . . ".blue().bold());
+    use crate::models::Config;
+    let mut config = Config::load().unwrap_or_default();
+    if args.value_of("cache").is_none() {
+        //attempt to fetch new themes
+        println!("    Retrieving themes from {:?}", &config.source());
+        let resp = reqwest::blocking::get(&config.source())?.text_with_charset("utf-8")?;
+        trace!("{:?}", &resp);
+
+        //compare to old themes
+        config.compare(toml::from_str(&resp)?)?;
+        trace!("{:?}", &config);
+        Config::save(&config)?;
+    }
+
+    //List themes
+    println!("{}", "\nAvailable themes:".blue().bold());
     for x in 0..config.theme.len() {
         let current = match config.theme[x].current {
             Some(true) => "Current: ".magenta().bold(),
             _ => "".white(),
         };
-        if config.theme[x].directory.is_some() {
-            println!(
-                "    Updating . . . {}{}: {}",
-                current,
-                config.theme[x].name,
-                config.theme[x]
-                    .description
-                    .as_ref()
-                    .unwrap_or(&"A LeftWM theme".to_string())
-            );
-            let repo = Repository::open(config.theme[x].directory.clone()?)?;
-            match fetch_origin_main(&repo) {
-                Ok(_) => {
-                    //if defined, attempt to checkout the specific index
-                    if config.theme[x].commit.is_some()
-                        && config.theme[x]
-                            .commit
-                            .clone()
-                            .unwrap_or_else(|| "".to_string())
-                            != *"*"
-                    {
-                        repo.set_head_detached(Oid::from_str(&config.theme[x].commit.as_ref()?)?)?;
-                        repo.checkout_head(None)?;
-                    }
-                }
-                Err(e) => {
-                    trace!("Error: {:?}", e);
-                    error!("Could not fetch repo.");
-                }
-            }
+        let installed = match config.theme[x].directory {
+            Some(_) => "-Installed".red().bold(),
+            None => "".white(),
+        };
+        println!(
+            "    {}{}: {}{}",
+            current,
+            config.theme[x].name,
+            config.theme[x]
+                .description
+                .as_ref()
+                .unwrap_or(&"A LeftWM theme".to_string()),
+            installed
+        );
+    }
 
-            installed += 1;
-        }
-    }
-    if installed == 0 {
-        println!("{}", "No themes installed.".red().bold());
-    }
     Ok(())
-}
-
-fn fetch_origin_main(repo: &git2::Repository) -> Result<(), git2::Error> {
-    repo.find_remote("origin")?.fetch(&["main"], None, None)
 }
